balena-image: Increase rootfs size to 1GB
The current limit is calculated so the whole image will take 700MB.
The nearly 300MB per rootfs partition is not enough.

The value of IMAGE_ROOTFS_MAXSIZE is set to IMAGE_ROOTFS_SIZE
rounded down to the nearest disk aligned size.
Align IMAGE_ROOTFS_SIZE or it will be larger than the limit.

Signed-off-by: Atanas Bunchev <atanas.bunchev@konsulko.com>